Consolidated ReloadingProfileCredentialsProvider functionality into ProfileCredentialsProvider
