+++ core/auth/src/main/java/software/amazon/awssdk/auth/credentials/ProfileCredentialsProvider.java
@@ -15,15 +15,23 @@
 
 package software.amazon.awssdk.auth.credentials;
 
+import java.nio.file.Path;
+import java.time.Clock;
+import java.time.Duration;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.function.Supplier;
 import software.amazon.awssdk.annotations.SdkPublicApi;
 import software.amazon.awssdk.annotations.SdkTestInternalApi;
 import software.amazon.awssdk.auth.credentials.internal.ProfileCredentialsUtils;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.profiles.ProfileFile;
+import software.amazon.awssdk.profiles.ProfileFileLocation;
 import software.amazon.awssdk.profiles.ProfileFileSystemSetting;
+import software.amazon.awssdk.profiles.internal.ProfileFileRefresher;
 import software.amazon.awssdk.utils.IoUtils;
 import software.amazon.awssdk.utils.SdkAutoCloseable;
 import software.amazon.awssdk.utils.ToString;
@@ -46,55 +54,73 @@ public final class ProfileCredentialsProvider
     implements AwsCredentialsProvider,
                SdkAutoCloseable,
                ToCopyableBuilder<ProfileCredentialsProvider.Builder, ProfileCredentialsProvider> {
-    private final AwsCredentialsProvider credentialsProvider;
+    private AwsCredentials credentials;
     private final RuntimeException loadException;
-
-    private final ProfileFile profileFile;
+    private final ProfileFileRefresher profileFileRefresher;
+    private final Supplier<ProfileFile> profileFileSupplier;
+    private final Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> reloadPredicate;
     private final String profileName;
-
     private final Supplier<ProfileFile> defaultProfileFileLoader;
+    private final Clock clock;
+    private final Duration refreshDuration;
+    private final Duration pollingDuration;
+    private final Duration prefetchTime;
 
     /**
      * @see #builder()
      */
     private ProfileCredentialsProvider(BuilderImpl builder) {
-        AwsCredentialsProvider credentialsProvider = null;
-        RuntimeException loadException = null;
-        ProfileFile profileFile = null;
-        String profileName = null;
+        this.defaultProfileFileLoader = builder.defaultProfileFileLoader;
+        this.reloadPredicate = builder.profileFileReloadPredicate;
+        this.clock = builder.clock;
+        this.refreshDuration = builder.refreshDuration;
+        this.pollingDuration = builder.pollingDuration;
+        this.prefetchTime = builder.prefetchTime;
+
+        RuntimeException thrownException = null;
+        String selectedProfileName = null;
+        Supplier<ProfileFile> selectedProfileSupplier = null;
+        ProfileFileRefresher profileRefresher = null;
 
         try {
-            profileName = builder.profileName != null ? builder.profileName
-                                                      : ProfileFileSystemSetting.AWS_PROFILE.getStringValueOrThrow();
+            selectedProfileName = Optional.ofNullable(builder.profileName)
+                                          .orElseGet(ProfileFileSystemSetting.AWS_PROFILE::getStringValueOrThrow);
 
-            // Load the profiles file
-            profileFile = Optional.ofNullable(builder.profileFile)
-                                  .orElseGet(builder.defaultProfileFileLoader);
+            selectedProfileSupplier = builder.profileFileSupplier;
+            Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> selectedPredicate = builder.profileFileReloadPredicate;
+            if (Objects.isNull(builder.profileFileSupplier)) {
+                selectedProfileSupplier = builder.defaultProfileFileLoader;
+                selectedPredicate = builder.defaultProfileFileReloadPredicate;
+            }
 
-            // Load the profile and credentials provider
-            String finalProfileName = profileName;
-            ProfileFile finalProfileFile = profileFile;
-            credentialsProvider =
-                    profileFile.profile(profileName)
-                               .flatMap(p -> new ProfileCredentialsUtils(finalProfileFile, p, finalProfileFile::profile)
-                                       .credentialsProvider())
-                               .orElseThrow(() -> {
-                                   String errorMessage = String.format("Profile file contained no credentials for " +
-                                                                       "profile '%s': %s", finalProfileName, finalProfileFile);
-                                   return SdkClientException.builder().message(errorMessage).build();
-                               });
+            if (Objects.nonNull(selectedProfileSupplier) && Objects.isNull(selectedPredicate)) {
+                thrownException = new IllegalArgumentException("Need to specify profileFileReloadPredicate when passing "
+                                                               + "profileFileSupplier");
+            }
+
+            profileRefresher = ProfileFileRefresher.builder()
+                                                   .profileFileSupplier(selectedProfileSupplier)
+                                                   .profileFileReloadPredicate(selectedPredicate)
+                                                   .exceptionHandler(builder.exceptionHandler)
+                                                   .addOnProfileFileReload(this::handleProfileFileReload)
+                                                   .refreshDuration(builder.refreshDuration)
+                                                   .pollingDuration(builder.pollingDuration)
+                                                   .prefetchTime(builder.prefetchTime)
+                                                   .clock(builder.clock)
+                                                   .asyncRefreshEnabled(builder.asyncRefreshEnabled)
+                                                   .build();
+
         } catch (RuntimeException e) {
             // If we couldn't load the credentials provider for some reason, save an exception describing why. This exception
             // will only be raised on calls to resolveCredentials. We don't want to raise an exception here because it may be
             // expected (eg. in the default credential chain).
-            loadException = e;
+            thrownException = e;
         }
 
-        this.loadException = loadException;
-        this.credentialsProvider = credentialsProvider;
-        this.profileFile = profileFile;
-        this.profileName = profileName;
-        this.defaultProfileFileLoader = builder.defaultProfileFileLoader;
+        this.loadException = thrownException;
+        this.profileName = selectedProfileName;
+        this.profileFileSupplier = selectedProfileSupplier;
+        this.profileFileRefresher = profileRefresher;
     }
 
     /**
@@ -127,27 +153,64 @@ public final class ProfileCredentialsProvider
         if (loadException != null) {
             throw loadException;
         }
-        return credentialsProvider.resolveCredentials();
+
+        refreshProfileFile();
+
+        return credentials;
+    }
+
+    private void handleProfileFileReload(ProfileFile profileFile) {
+        updateCredentials(profileFile, profileName);
+    }
+
+    private void updateCredentials(ProfileFile profileFile, String profileName) {
+        AwsCredentialsProvider credentialsProvider = createCredentialsProvider(profileFile, profileName);
+
+        this.credentials = credentialsProvider.resolveCredentials();
+
+        IoUtils.closeIfCloseable(credentialsProvider, null);
+    }
+
+    private ProfileFile refreshProfileFile() {
+        return profileFileRefresher.refreshIfStale();
     }
 
     @Override
     public String toString() {
         return ToString.builder("ProfileCredentialsProvider")
                        .add("profileName", profileName)
-                       .add("profileFile", profileFile)
+                       .add("profileFileSupplier", profileFileSupplier)
                        .build();
     }
 
     @Override
     public void close() {
-        // The delegate credentials provider may be closeable (eg. if it's an STS credentials provider). In this case, we should
-        // clean it up when this credentials provider is closed.
-        IoUtils.closeIfCloseable(credentialsProvider, null);
+        profileFileRefresher.close();
     }
 
     @Override
     public Builder toBuilder() {
         return new BuilderImpl(this);
+    }
+
+    private AwsCredentialsProvider createCredentialsProvider(ProfileFile profileFile, String profileName) {
+        // Load the profile and credentials provider
+        return profileFile.profile(profileName)
+                          .flatMap(p -> new ProfileCredentialsUtils(profileFile, p, profileFile::profile).credentialsProvider())
+                          .orElseThrow(() -> {
+                              String errorMessage = String.format("Profile file contained no credentials for " +
+                                                                  "profile '%s': %s", profileName, profileFile);
+                              return SdkClientException.builder().message(errorMessage).build();
+                          });
+    }
+
+    /**
+     * Convenience method for creating a Predicate function to pass to {@link Builder#profileFileReloadPredicate(Predicate)}.
+     * @param path The credentials file path.
+     * @return A Predicate instance for testing if the ProfileFile was refreshed before the last modified timestamp of the file.
+     */
+    public static Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> wasRefreshedBeforeFileModificationTime(Path path) {
+        return refreshRecord -> refreshRecord.wasCreatedBeforeFileModified(path);
     }
 
     /**
@@ -158,6 +221,8 @@ public final class ProfileCredentialsProvider
         /**
          * Define the profile file that should be used by this credentials provider. By default, the
          * {@link ProfileFile#defaultProfileFile()} is used.
+         * This method does not provide reloading functionality as {@link #profileFileSupplier} and
+         * {@link #profileFileReloadPredicate(Predicate)} do.
          */
         Builder profileFile(ProfileFile profileFile);
 
@@ -174,6 +239,61 @@ public final class ProfileCredentialsProvider
         Builder profileName(String profileName);
 
         /**
+         * Define the mechanism for loading profile files.
+         *
+         * @param profileFileSupplier Supplier for generating a ProfileFile instance.
+         */
+        Builder profileFileSupplier(Supplier<ProfileFile> profileFileSupplier);
+
+        /**
+         * Define the condition for reloading a profile file.
+         *
+         * @param predicate Predicate for determining whether to execute the profileFileSupplier.
+         */
+        Builder profileFileReloadPredicate(Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> predicate);
+
+        /**
+         * Define the frequency with which the credentials provider will check whether the profile file may have outstanding
+         * changes and needs to be reloaded.
+         *
+         * <p>The automatic refreshing of the ProfileFile can be disabled by setting value to null</p>
+         *
+         * @param refreshDuration Once credentials are loaded, the credentials provider will suspend refreshing and
+         *                        only restart until this period has elapsed.
+         */
+        Builder refreshDuration(Duration refreshDuration);
+
+        /**
+         * Define the frequency with which the credentials provider will check whether the profile file may have outstanding
+         * changes and needs to be reloaded.
+         *
+         * @param pollingDuration Once the credentials provider has started refreshing, the profile file will be checked
+         *                        for changes with this frequency.
+         */
+        Builder pollingDuration(Duration pollingDuration);
+
+        /**
+         * Configure the amount of time, relative to refresh and polling time, that the cached profile file is considered
+         * close to stale and should be updated. See {@link #asyncRefreshEnabled}.
+         */
+        Builder prefetchTime(Duration prefetchTime);
+
+        /**
+         * @param exceptionHandler Handler which takes action when a Runtime exception occurs while loading a profile file.
+         *                         Handler can return a previously stored profile file or throw back the exception.
+         */
+        Builder exceptionHandler(Function<RuntimeException, ProfileFile> exceptionHandler);
+
+        /**
+         * Configure whether this refresher should fetch tokens asynchronously in the background. If this is true, threads are
+         * less likely to block when {@link #resolveCredentials()} is called, but additional resources are used to maintain the
+         * provider.
+         *
+         * <p>By default, this is disabled.</p>
+         */
+        Builder asyncRefreshEnabled(Boolean asyncRefreshEnabled);
+
+        /**
          * Create a {@link ProfileCredentialsProvider} using the configuration applied to this builder.
          */
         @Override
@@ -181,23 +301,38 @@ public final class ProfileCredentialsProvider
     }
 
     static final class BuilderImpl implements Builder {
-        private ProfileFile profileFile;
+        private Supplier<ProfileFile> profileFileSupplier;
+        private Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> profileFileReloadPredicate;
         private String profileName;
         private Supplier<ProfileFile> defaultProfileFileLoader = ProfileFile::defaultProfileFile;
+        private Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> defaultProfileFileReloadPredicate
+            = refreshRecord -> refreshRecord.wasCreatedBeforeFileModified(ProfileFileLocation.credentialsFilePath());
+        private Function<RuntimeException, ProfileFile> exceptionHandler = exceptionHandler();
+        private Clock clock = Clock.systemUTC();
+        private Duration refreshDuration;
+        private Duration pollingDuration;
+        private Duration prefetchTime;
+        private Boolean asyncRefreshEnabled = Boolean.FALSE;
 
         BuilderImpl() {
         }
 
         BuilderImpl(ProfileCredentialsProvider provider) {
-            this.profileFile = provider.profileFile;
             this.profileName = provider.profileName;
             this.defaultProfileFileLoader = provider.defaultProfileFileLoader;
+            this.profileFileSupplier = provider.profileFileSupplier;
+            this.profileFileReloadPredicate = provider.reloadPredicate;
+            this.clock = provider.clock;
+            this.refreshDuration = provider.refreshDuration;
+            this.pollingDuration = provider.pollingDuration;
+            this.prefetchTime = provider.prefetchTime;
         }
 
         @Override
         public Builder profileFile(ProfileFile profileFile) {
-            this.profileFile = profileFile;
-            return this;
+            Supplier<ProfileFile> supplier = Objects.nonNull(profileFile) ? () -> profileFile : null;
+
+            return profileFileSupplier(supplier).profileFileReloadPredicate(r -> false);
         }
 
         public void setProfileFile(ProfileFile profileFile) {
@@ -220,19 +355,112 @@ public final class ProfileCredentialsProvider
         }
 
         @Override
+        public Builder profileFileSupplier(Supplier<ProfileFile> profileFileSupplier) {
+            this.profileFileSupplier = profileFileSupplier;
+            return this;
+        }
+
+        public void setProfileFileSupplier(Supplier<ProfileFile> supplier) {
+            profileFileSupplier(supplier);
+        }
+
+        @Override
+        public Builder profileFileReloadPredicate(Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> predicate) {
+            this.profileFileReloadPredicate = predicate;
+            return this;
+        }
+
+        public void setProfileFileReloadPredicate(Predicate<ProfileFileRefresher.ProfileFileRefreshRecord> predicate) {
+            profileFileReloadPredicate(predicate);
+        }
+
+        Builder clock(Clock clock) {
+            this.clock = clock;
+            return this;
+        }
+
+        @Override
+        public Builder refreshDuration(Duration refreshDuration) {
+            this.refreshDuration = refreshDuration;
+            return this;
+        }
+
+        public void setRefreshDuration(Duration refreshDuration) {
+            refreshDuration(refreshDuration);
+        }
+
+        @Override
+        public Builder pollingDuration(Duration pollingDuration) {
+            this.pollingDuration = pollingDuration;
+            return this;
+        }
+
+        public void setPollingDuration(Duration pollingDuration) {
+            refreshDuration(pollingDuration);
+        }
+
+        @Override
+        public Builder prefetchTime(Duration prefetchTime) {
+            this.prefetchTime = prefetchTime;
+            return this;
+        }
+
+        public void setPrefetchTime(Duration prefetchTime) {
+            prefetchTime(prefetchTime);
+        }
+
+        @Override
+        public Builder exceptionHandler(Function<RuntimeException, ProfileFile> exceptionHandler) {
+            this.exceptionHandler = exceptionHandler;
+            return this;
+        }
+
+        public void setExceptionHandler(Function<RuntimeException, ProfileFile> exceptionHandler) {
+            exceptionHandler(exceptionHandler);
+        }
+
+        @Override
+        public Builder asyncRefreshEnabled(Boolean asyncRefreshEnabled) {
+            this.asyncRefreshEnabled = asyncRefreshEnabled;
+            return this;
+        }
+
+        public void setAsyncRefreshEnabled(Boolean asyncRefreshEnabled) {
+            asyncRefreshEnabled(asyncRefreshEnabled);
+        }
+
+        @Override
         public ProfileCredentialsProvider build() {
             return new ProfileCredentialsProvider(this);
         }
 
         /**
          * Override the default configuration file to be used when the customer does not explicitly set
-         * profileName(profileName);
-         * {@link #profileFile(ProfileFile)}. Use of this method is only useful for testing the default behavior.
+         * profileFile(ProfileFile) or profileFileSupplier(supplier);
+         * {@link #profileFile(ProfileFile)}, {@link #profileFileSupplier(Supplier)}. Use of this method is
+         * only useful for testing the default behavior.
          */
         @SdkTestInternalApi
         Builder defaultProfileFileLoader(Supplier<ProfileFile> defaultProfileFileLoader) {
             this.defaultProfileFileLoader = defaultProfileFileLoader;
             return this;
         }
+
+        /**
+         * Override the location of the default configuration file to be used when the customer does not explicitly set
+         * profileFile(ProfileFile) or profileFileSupplier(supplier);
+         * {@link #profileFileReloadPredicate(Predicate)}. Use of this method is only useful for testing the default behavior.
+         */
+        @SdkTestInternalApi
+        Builder defaultProfileFileReloadPredicate(Predicate<ProfileFileRefresher.ProfileFileRefreshRecord>
+                                                      defaultProfileFileReloadPredicate) {
+            this.defaultProfileFileReloadPredicate = defaultProfileFileReloadPredicate;
+            return this;
+        }
+
+        private static Function<RuntimeException, ProfileFile> exceptionHandler() {
+            return e -> { throw e; };
+        }
     }
+
 }
